refactor(lesson-deep-dive): move the challenge and gallery wire types into the package
The gallery and challenge activities validate the same server payloads and their components are about to move into the package; the shared types and validators go first so the later moves never reach back into apps.
