+++ apps/src/aiTutor/views/gallery/AssessmentPanel.tsx
@@ -1,11 +1,10 @@
 import FontAwesomeV6Icon from '@code-dot-org/component-library/fontAwesomeV6Icon';
+import {ChallengeResponseDetail} from '@code-dot-org/lesson-deep-dive';
 import {Typography} from '@mui/material';
 import classNames from 'classnames';
 import React, {FC} from 'react';
 
 import aiTutorAvatar from '@cdo/static/tutor/ai-tutor-avatar.png';
-
-import {ChallengeResponseDetail} from './types';
 
 import styles from './project-view.module.scss';
