+++ apps/src/aiTutor/views/gallery/ChallengeGallery.tsx
@@ -1,17 +1,18 @@
+import {
+  ChallengeResponse,
+  challengeResponseListValidator,
+  GallerySort,
+  Reaction,
+  TutorGalleryData,
+  unitCountsValidator,
+} from '@code-dot-org/lesson-deep-dive';
 import React, {FC, useEffect, useState} from 'react';
 
 import HttpClient from '@cdo/apps/util/HttpClient';
 
-import {
-  ChallengeResponse,
-  Reaction,
-  challengeResponseListValidator,
-} from '../lessonDeepDive/types';
-
 import GallerySidebar from './GallerySidebar';
 import ProjectCard, {ProjectVariant} from './ProjectCard';
 import ProjectView from './ProjectView';
-import {GallerySort, TutorGalleryData, unitCountsValidator} from './types';
 
 import styles from './challenge-gallery.module.scss';
