+++ apps/src/aiTutor/views/gallery/ProjectDetailsCard.tsx
@@ -1,12 +1,13 @@
+import {
+  assetWithUrl,
+  ChallengeResponseDetail,
+  Reaction,
+} from '@code-dot-org/lesson-deep-dive';
 import {Typography} from '@mui/material';
 import classNames from 'classnames';
 import React, {FC} from 'react';
 
-import {Reaction} from '../lessonDeepDive/types';
-
-import {assetWithUrl} from './assetUtils';
 import ReactionChips from './ReactionChips';
-import {ChallengeResponseDetail} from './types';
 
 import styles from './project-view.module.scss';
