+++ apps/src/aiTutor/views/lessonDeepDive/ChallengeActivities/VideoChallenge.tsx
@@ -1,16 +1,15 @@
-import {VideoRecorder} from '@code-dot-org/lesson-deep-dive';
+import {
+  VideoRecorder,
+  Challenge,
+  ChallengeResponse,
+  challengeResponseValidator,
+} from '@code-dot-org/lesson-deep-dive';
 import React, {FC, useCallback, useEffect, useRef, useState} from 'react';
 
 import AichatContextManager from '@cdo/apps/aichat/aichatContextManager';
 import {getClientApi} from '@cdo/apps/aichat/api/client';
 import HttpClient from '@cdo/apps/util/HttpClient';
 import {AiChatClientTypes} from '@cdo/generated-scripts/sharedConstants';
-
-import {
-  Challenge,
-  ChallengeResponse,
-  challengeResponseValidator,
-} from '../types';
 
 import {requestEvaluation} from './requestEvaluation';
