+++ apps/src/aiTutor/views/lessonDeepDive/ChallengeActivities/WhiteboardChallenge.tsx
@@ -2,6 +2,7 @@ import {
   useTheme,
   ThemeProvider,
 } from '@code-dot-org/component-library/common/contexts';
+import {ExplanationTypes} from '@code-dot-org/lesson-deep-dive';
 import {createTheme, ThemeProvider as MuiThemeProvider} from '@mui/material';
 import {ReactFlowProvider, useReactFlow} from '@xyflow/react';
 import '@xyflow/react/dist/style.css';
@@ -16,8 +17,6 @@ import {createSketchSnapshotBlob} from '@cdo/apps/sketchlab/reactFlow/utils/crea
 import HttpClient from '@cdo/apps/util/HttpClient';
 import {createUuid} from '@cdo/apps/utils';
 import {AiChatClientTypes} from '@cdo/generated-scripts/sharedConstants';
-
-import {ExplanationTypes} from '../types';
 
 import AudioRecorder from './AudioRecorder';
 import {requestEvaluation} from './requestEvaluation';
