+++ apps/test/unit/aiTutor/views/gallery/ChallengeGalleryTest.tsx
@@ -1,14 +1,14 @@
+import {
+  ChallengeResponse,
+  ChallengeResponseDetail,
+  TutorGalleryData,
+} from '@code-dot-org/lesson-deep-dive';
 import {fireEvent, render, screen, waitFor} from '@testing-library/react';
 import '@testing-library/jest-dom';
 import React from 'react';
 
 import ChallengeGallery from '@cdo/apps/aiTutor/views/gallery/ChallengeGallery';
 import {addReaction} from '@cdo/apps/aiTutor/views/gallery/reactionsApi';
-import {
-  ChallengeResponseDetail,
-  TutorGalleryData,
-} from '@cdo/apps/aiTutor/views/gallery/types';
-import {ChallengeResponse} from '@cdo/apps/aiTutor/views/lessonDeepDive/types';
 import HttpClient from '@cdo/apps/util/HttpClient';
 
 jest.mock('@cdo/apps/util/HttpClient', () => ({
